pyproject.toml
[project]
name = "chess-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "python-chess>=1.999",
    "stockfish>=3.28.0",
    "uvicorn>=0.34.2",
]
